+++ src/names.h
@@ -109,6 +109,9 @@ tree_t find_std(nametab_t *tab);
 tree_t find_forward_decl(nametab_t *tab, tree_t decl);
 psl_node_t find_default_clock(nametab_t *tab);
 
+typedef void (*predef_cb_t)(tree_t, void *);
+void walk_predefs(nametab_t *tab, ident_t name, predef_cb_t fn, void *context);
+
 type_t solve_types(nametab_t *tab, tree_t expr, type_t constraint);
 type_t solve_known_subtype(nametab_t *tab, tree_t expr, type_t constraint);
 type_t solve_target(nametab_t *tab, tree_t target, tree_t value);