+++ test/test_parse.c
@@ -4799,6 +4799,8 @@ END_TEST
 
 START_TEST(test_visibility3)
 {
+   set_standard(STD_08);
+
    input_from_file(TESTDIR "/parse/visibility3.vhd");
 
    for (int i = 0; i < 2; i++) {
@@ -5785,6 +5787,56 @@ START_TEST(test_issue761)
 }
 END_TEST
 
+START_TEST(test_issue776)
+{
+   set_standard(STD_08);
+
+   input_from_file(TESTDIR "/parse/issue776.vhd");
+
+   tree_t p = parse();
+   fail_if(p == NULL);
+   fail_unless(tree_kind(p) == T_PACKAGE);
+   lib_put(lib_work(), p);
+
+   tree_t e = parse();
+   fail_if(e == NULL);
+   fail_unless(tree_kind(e) == T_ENTITY);
+   lib_put(lib_work(), e);
+
+   tree_t a = parse();
+   fail_if(a == NULL);
+   fail_unless(tree_kind(a) == T_ARCH);
+
+   fail_unless(parse() == NULL);
+
+   fail_if_errors();
+}
+END_TEST
+
+START_TEST(test_visibility10)
+{
+   input_from_file(TESTDIR "/parse/visibility10.vhd");
+
+   tree_t p1 = parse();
+   fail_if(p1 == NULL);
+   fail_unless(tree_kind(p1) == T_PACKAGE);
+   lib_put(lib_work(), p1);
+
+   tree_t p2 = parse();
+   fail_if(p2 == NULL);
+   fail_unless(tree_kind(p2) == T_PACKAGE);
+   lib_put(lib_work(), p2);
+
+   tree_t p3 = parse();
+   fail_if(p3 == NULL);
+   fail_unless(tree_kind(p3) == T_PACKAGE);
+
+   fail_unless(parse() == NULL);
+
+   fail_if_errors();
+}
+END_TEST
+
 Suite *get_parse_tests(void)
 {
    Suite *s = suite_create("parse");
@@ -5908,6 +5960,8 @@ Suite *get_parse_tests(void)
    tcase_add_test(tc_core, test_issue753);
    tcase_add_test(tc_core, test_issue760);
    tcase_add_test(tc_core, test_issue761);
+   tcase_add_test(tc_core, test_issue776);
+   tcase_add_test(tc_core, test_visibility10);
    suite_add_tcase(s, tc_core);
 
    return s;