+++ NEWS.md
@@ -25,6 +25,8 @@
 - Support for type conversions between arrays with closely related
   element types.
 - Added support for FSM state coverage collection (from @Blebowski).
+- An alias of a type now correctly creates implicit aliases for each
+  predefined operator of that type (#776).
 
 ## Version 1.10.4 - 2023-10-16
 - Fixed compatibility with LLVM 17.